finished the project

+++ ft_printf/src/print_hex.c
@@ -6,13 +6,13 @@
 /*   By: sabdulqa <sabdulqa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/11 15:35:53 by sabdulqa          #+#    #+#             */
-/*   Updated: 2023/02/26 17:07:06 by sabdulqa         ###   ########.fr       */
+/*   Updated: 2023/02/28 15:34:50 by sabdulqa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-int	print_hex(unsigned int n, char type)
+int	print_hex(unsigned long long n, char type)
 {
 	char	*digits;
 	int		count;
